:memo: Slightly improve paste performance
We can’t really fix how the spellchecker works, but we can fix *when*
it works. This is a poor man’s solution to our original problem (pastes
taking too much time) - whenever text is pasted into the textarea, we
briefly disable the spellchecker, just to re-enable it a few moments
later.

Closes #168